wip: add size_t version of fib generator
Note that the index is still int, but the results are std::size_t.
There is issues using std::size_t as index due to the fib calculation
f(N - 1) + f(N - 2), there is a chance of initiating f(-1) which will be
come a super huge number with std::size_t.